+++ example/BUILD.bazel
@@ -6,8 +6,8 @@ go_library(
     importpath = "github.com/uber/tango/example",
     visibility = ["//visibility:private"],
     deps = [
+        "//config",
         "//controller",
-        "//core/config",
         "//core/git",
         "//core/repomanager",
         "//core/storage",
