+++ graphrunner/BUILD.bazel
@@ -10,8 +10,8 @@ go_library(
     importpath = "github.com/uber/tango/graphrunner",
     visibility = ["//visibility:public"],
     deps = [
+        "//config",
         "//core/bazel",
-        "//core/config",
         "//core/git",
         "//core/targethasher",
         "//core/workspace",
